Initial import of the graphics core
This is intended to become a framebuffer-based graphics controller
producing 720p output.

This is derived from an older similar design in my earlier 68k-computer
project, although this new adaptation of it isn't quite working yet
because I've not fully tailored it for the BeagleWire development board
I'm intending to use this time. (The previous iteration was developed
on an ICEBreaker development board, but BeagleWire is more suitable this
time because it has SDRAM on board already, reducing the number of
things we need to connect.