graphics: main module external bus interfaces
These are how the main graphics core will interact with the host CPU and
with the video RAM. The intention here is that the main core treats these
as generic asynchronous memory buses and then the board-specific caller
will adapt the requests to suit whatever the physical devices actually
expect. For example, the caller will likely connect the video RAM bus to
an SDRAM controller.