graphics: new timing signals for pipeline preparation
As we move towards building the display from video RAM rather than from
hard-coded logic, we'll need some advance notice from our timing logic
that a frame and/or line is about to start so we can start a pipeline of
requesting video data just before we need it.

To achieve that, we designate the back porches of both the horizontal and
vertical blanking intervals as "preparation time". During that phase, the
X and Y coordinates count up from a negative number of pixels towards
zero, with zero marking the first pixel or line of the active display.

We also include new output signals "hprep" and "vprep" which allow
synchronized callers to recognize when the preparation phase is active so
they can watch for hitting the appropriate number of clocks before the
active display starts and beginning whatever requests they need to make.

The counters still behave as before during the active display states, so
simple callers like our test pattern module (which doesn't need to do any
data fetching) can just ignore the preparation phase and work directly
with the active pixel counters as before.