graphics: dual_port_buffer module for FIFO block RAM
To make it more likely that the async FIFO's buffer will be inferred as
a block RAM, we need to read the output always into the same register,
which we can then pass on to the caller.

Our buffer here works on the negative edges of the respective clocks to
avoid adding a whole cycle of lag as a result of now reading through an
additional register.