graphics: Refactor around an async FIFO pattern
Previously we had the display timing feeding directly into the test
pattern, which in turn drove the pixel data outputs. That worked okay
for the test pattern, but isn't suitable for other data sources like an
SDRAM where the pixel data rate varies due to write operations and refresh
operations.

The new model is to have the data source fill a FIFO queue as fast as it
can, and then a pixel from the FIFO for each visible pixel. That then
decouples the data fetching rate from the pixel rate, as long as the
data fetching is able to stay ahead of the raster.

This is a first pass of the necessary refactoring to get the desired
behavior, but it needs considerable additional work to close timing.
Firstly, the definition of the FIFO buffer is not yet being inferred as
a block RAM, so we have lots of delay from storing that data in logic
cells. Also, the test pattern logic is a bit heavy for the 100MHz clock
we're going to be using for data (since that's what our SDRAM expects)
and so we may need to simplify it to get a working result.